requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
openpyxl
python-multipart
bcrypt
PyJWT
google-cloud-secret-manager
websockets
slowapi
google-cloud-bigquery
